Introduce `Frame`
A `Frame` encapsulates `Header` and `Data`. Currently only `Event`
frames are supported. All other frame types are not yet implemented.

The `App` has been updated to iterate of `Frame`s instead of `Header`s.

Lastly, a new error type has been introduced, called `AppError`. This
type is going to the standard type for binary users of this crate. While
`ParseError` is an error intended for users of the library.