Reorganize project
All of the types that build up a `Frame` and their parsers were in
`src/lib.rs`. This commit splits the parsers from the data types.  The
data types are also split up. The `Header` and a few related types are
in `src/frames/mod.rs`. The data structures related to the event frame
are in `src/frames/event.rs`. All other frame types should be in it's
own module.