Add script to plot race line as SVG
The script plots the race line of the human driver as SVG. I used this
script to explore the design of a high level API.

The F1 2020 API exposes a lot of info using 10 different type of frames.
Currently, this API has been partly implemented by this crate. I call
this the low-level API.

I'm trying to come up with a high-level API that is easier to use than
the low-level API. The results of that effort can be found in
src/race.rs.